chore(arena): split arena db services
