+++ backend/arena/services/comparison.py
@@ -1,10 +1,7 @@
 import uuid
 from datetime import datetime
-from typing import TYPE_CHECKING, TypeVar
 
-from fastapi import HTTPException, status
-from linkup import LinkupSearchTextResult
-
+from backend.arena.services.utils import get_item
 from backend.llms.data import get_llms_data
 from utils.database.models import (
     BOT_POS,
@@ -13,35 +10,9 @@ from utils.database.models import (
     ComparisonCreate,
     ComparisonRead,
     ErrorDetails,
-    LLMMessage,
-    LLMMessageCreate,
     SystemMessage,
-    Turn,
-    TurnCreate,
-    TurnRead,
-    TurnVoteAnnotate,
-    TurnVoteChoice,
-    UserMessage,
 )
 from utils.database.session import get_session
-
-if TYPE_CHECKING:
-    from sqlmodel.ext.asyncio.session import AsyncSession
-
-
-T = TypeVar("T", bound=Comparison | Turn)
-
-
-async def _get_item(item_class: type[T], id: uuid.UUID, session: "AsyncSession") -> T:
-    db_item = await session.get(item_class, id)
-
-    if not db_item:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail=f"{item_class.__name__} not found",
-        )
-
-    return db_item
 
 
 async def create_comparison(comparison: ComparisonCreate) -> ComparisonRead:
@@ -64,7 +35,7 @@ async def create_comparison(comparison: ComparisonCreate) -> ComparisonRead:
 
 async def read_comparison(id: uuid.UUID) -> ComparisonRead:
     async with get_session() as session:
-        db_comparison = await _get_item(Comparison, id, session)
+        db_comparison = await get_item(Comparison, id, session)
 
     return ComparisonRead.model_validate(db_comparison)
 
@@ -100,7 +71,7 @@ async def update_comparison_llm_id(
             await session.delete(failing_system_msg)
             await session.flush()
 
-        db_comparison = await _get_item(Comparison, comparison.id, session)
+        db_comparison = await get_item(Comparison, comparison.id, session)
         setattr(db_comparison, f"llm_id_{pos}", new_llm_id)
         setattr(db_comparison, f"system_msg_{pos}", system_msg)
         db_comparison.custom_models_selection = comparison.custom_models_selection
@@ -115,61 +86,7 @@ async def update_comparison_error(
     comparison.error = error
 
     async with get_session() as session:
-        db_comparison = await _get_item(Comparison, comparison.id, session)
+        db_comparison = await get_item(Comparison, comparison.id, session)
         db_comparison.error = error.model_dump() if error else None
         db_comparison.updated_at = datetime.now()
-        await session.commit()
-
-
-async def add_comparison_turn(
-    comparison_id: uuid.UUID,
-    prompt: str,
-    web_search_results: list[LinkupSearchTextResult] | None = None,
-) -> tuple[ComparisonRead, TurnRead]:
-    async with get_session() as session:
-        db_turn = Turn.model_validate(
-            TurnCreate(
-                comparison_id=comparison_id,
-                user_msg=UserMessage(
-                    content=prompt,
-                    web_search_results=(
-                        [w.model_dump() for w in web_search_results]
-                        if web_search_results
-                        else None
-                    ),
-                ),
-            )
-        )
-        new_turn_id = db_turn.id
-        session.add(db_turn)
-        await session.commit()
-
-    comparison = await read_comparison(comparison_id)
-
-    return (comparison, next(t for t in comparison.turns if t.id == new_turn_id))
-
-
-async def update_turn(
-    id: uuid.UUID, llm_msg_a: LLMMessageCreate, llm_msg_b: LLMMessageCreate
-) -> None:
-    async with get_session() as session:
-        db_turn = await _get_item(Turn, id, session)
-        db_turn.llm_msg_a = LLMMessage.model_validate(llm_msg_a)
-        db_turn.llm_msg_b = LLMMessage.model_validate(llm_msg_b)
-        session.add(db_turn)
-        await session.commit()
-
-
-async def update_turn_vote(
-    id: uuid.UUID, vote: TurnVoteChoice | TurnVoteAnnotate
-) -> None:
-    async with get_session() as session:
-        db_turn = await _get_item(Turn, id, session)
-        data = (
-            {f"{k}_{vote.pos}": v for k, v in vote.model_dump().items()}
-            if isinstance(vote, TurnVoteAnnotate)
-            else vote.model_dump()
-        )
-        db_turn.sqlmodel_update(data)
-        session.add(db_turn)
         await session.commit()
